Add interface file and overloaded functions
Add an interface file to act as a "map" to easily see all the functions
we currently have defined, as well as to contain several enums and
constants used globally.

Overload RooFitter to take either a file name, or a file handle. The
latter is useful when running interactively in root with opened files
(_file0, _file1, etc.).

Also clean up includes to just those needed to compile.